unified space and spaces resolvers
